requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
deepface==0.0.79
pillow==10.0.0 